backend/requirements.txt
# -------------------------------
# Core framework
# -------------------------------
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
anyio==4.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1

# -------------------------------
# Audio / Video processing
# -------------------------------
ffmpeg-python==0.2.0
av==15.0.0
faster-whisper==1.2.0
ctranslate2==4.6.0
onnxruntime==1.22.1
numpy==2.3.2
pillow>=10.3.0
opencv-python-headless>=4.10.0.84

# -------------------------------
# OCR (legacy + helpers)
# -------------------------------
pytesseract>=0.3.10
rapidfuzz>=3.9.5

# -------------------------------
# PaddleOCR stack
# NOTE: Install PaddlePaddle separately on Apple Silicon:
#   pip install --upgrade "paddlepaddle==2.6.1" --extra-index-url https://www.paddlepaddle.org.cn/whl/macos/mps
# Then install paddleocr (no extra heavy PDF deps needed for our use):
# -------------------------------
paddleocr==2.7.0.3
shapely>=2.0.4
pyclipper>=1.3.0.post5
scikit-image>=0.24.0
fire>=0.6.0
python-Levenshtein>=0.26.0
lmdb>=1.4.1

# -------------------------------
# Logging & utilities
# -------------------------------
coloredlogs==15.0.1
humanfriendly==10.0
tqdm==4.67.1
requests==2.32.5
filelock==3.19.1
packaging==25.0
flatbuffers==25.2.10
future==1.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
protobuf==6.32.0
PyYAML==6.0.2
fsspec==2025.7.0
huggingface-hub==0.34.4
hf-xet==1.1.8

# -------------------------------
# Text-to-speech and helpers
# -------------------------------
pyttsx3
aiofiles

# -------------------------------
# Dev / Formatting / Testing
# -------------------------------
black
ruff
pytest
pre-commit
